requirements.txt
qrcode~=7.3.1
openpyxl~=3.0.10
Faker~=14.2.1
Pillow~=8.4.0
PyDrive~=1.3.1